Split the implementation of the roi item module
- base / common rois / arc roi
